requirements.txt
Flask==2.1.2
gunicorn==20.1.0
matplotlib==3.5.1
numpy==1.21.4
tensorflow-cpu==2.9.1
opencv-python-headless==4.6.0.66
trio-websocket==0.9.2